deepfake web app/requirements.txt
flask
flask-cors
torch
torchvision
opencv-python
pillow
